1.加入登录时拦截Cookie 2.我喜欢的音乐歌单

+++ app/src/main/java/com/fr1014/mycoludmusic/data/entity/http/wangyiyun/RecommendPlayList.kt
@@ -15,7 +15,7 @@ data class PlayListResult(
     val id: Long,
     val name: String,
     val picUrl: String,
-    val playCount: Int,
+    val playCount: Long,
     val trackCount: Int,
     val trackNumberUpdateTime: Long,
     val type: Int
